refactor: reorganize utils module into a package with dedicated submodules #77
